pyproject.toml
[project]
name = "crit"
version = "0.1.0"
description = ""
authors = [
    {name = "Duncan Brown",email = "duncanjbrown@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "streamlit (>=1.43.0,<2.0.0)",
    "markdownify (>=1.1.0,<2.0.0)",
    "openai (>=1.65.4,<2.0.0)",
    "langchain (>=0.3.20,<0.4.0)",
    "langchain-openai (>=0.3.7,<0.4.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "mistune (>=3.1.2,<4.0.0)",
    "tenacity (>=9.0.0,<10.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
